Implement Load Phase.
Load phase is added to load data in new components without complex preconditioning.

All ClearStats functions are used to erase Stats recorded in load phase.
We can check only the Stats recorded in run phase.

If you want to use the load phase to load data in all components, all you need to do is adding the Load_File_Path attribute in the workload xml file.
If you don't want to use the load phase to load data in the components, you just modify the value of Load_file_Path attribute as "NONE".